# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "This is the documentation for our Mqtt server api",
      "version": "1.0.0",
      "title": "OBD Server API"
    },
    "host": "localhost:2000",   
    "produces": ["application/json"],
    "schemes": [
      "http"
    ],
    "paths":{
        "/api/registration": {
            "get": {
              "description": "Returns all users registred with our System",
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "Successfully got users",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                }
              },
              "parameters": [
                {
                  "in": "query",
                  "name": "name",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by name"
                },
                {
                  "in": "query",
                  "name": "phone",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by phone number"
                },
                {
                  "in": "query",
                  "name": "mail",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by mail"
                },
                {
                  "in": "query",
                  "name": "address",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by address"
                },
                {
                  "in": "query",
                  "name": "rfid",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by rfid"
                },
                {
                  "in": "query",
                  "name": "reg_num",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by Registration number aka topic"
                },
                {
                  "in": "query",
                  "name": "vin",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by vin"
                },
                {
                  "in": "query",
                  "name": "mac",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by mac"
                },
                {
                  "in": "query",
                  "name": "vm",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by vehicle Make"
                },
                {
                  "in": "query",
                  "name": "ft",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by fuel type"
                },
                {
                  "in": "query",
                  "name": "isobd",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by Is Obd"
                },
                {
                  "in": "query",
                  "name": "dc",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by date of creation"
                },
                {
                  "in": "query",
                  "name": "du",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by date of updation"
                },
                {
                  "in": "query",
                  "name": "fl",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by fuel level"
                },
                {
                  "in": "query",
                  "name": "tc",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by Tank capacity"
                },
                {
                  "in": "query",
                  "name": "rpm",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by RPM"
                },
                {
                  "in": "query",
                  "name": "speed",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by speed"
                },
                {
                  "in": "query",
                  "name": "maf",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by mass air flow"
                },
                {
                  "in": "query",
                  "name": "ctemp",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by coolant temperature"
                },
                {
                  "in": "query",
                  "name": "milage",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by milage"
                },
                {
                  "in": "query",
                  "name": "od1",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by Other data 1"
                },
                {
                  "in": "query",
                  "name": "od2",
                  "schema": {
                    "type": "string"
                  },
                  "required":false,
                  "description": "Filter by Other data 2"
                }
              ]
            },
            "post": {
              "description": "Post a user to the DB",
              "produces": [
                "application/json"
              ],
              "requestBody":{
                "description": "Enter user details",
                "content": {
                    "application/json":
                    {
                        "schema": {
                            "type": "object",
                            "properties": {
                              "CustomerName": {
                                "type": "string"
                              },
                              "PHONE":{
                                "type": "string"
                              },
                              "EMAIL":{
                                "type": "string"
                              },
                              "ADDRESS":{
                                "type": "string"
                              },
                              "RFID":{
                                "type": "string"
                              },
                              "REG_NUM":{
                                "type": "string"
                              },
                              "VIN":{
                                "type": "string"
                              },
                              "MAC":{
                                "type": "string"
                              },
                              "VEHICLE_MAKE":{
                                "type": "string"
                              },
                              "FUEL_TYPE":{
                                "type": "string"
                              },
                              "IS_OBD":{
                                "type": "string"
                              },
                              "DateOfCreation":{
                                "type": "string"
                              },
                              "DateOfUpdation":{
                                "type": "string"
                              },
                              "FUEL_LEVEL":{
                                "type": "string"
                              },
                              "TANK_CAPACITY":{
                                "type": "string"
                              },
                              "RPM":{
                                "type": "string"
                              },
                              "SPEED":{
                                "type": "string"
                              },
                              "MAF":{
                                "type": "string"
                              },
                              "COOLANT_TEMP":{
                                "type": "string"
                              },
                              "MILAGE":{
                                "type": "string"
                              },
                              "OtherData1":{
                                "type": "string"
                              },
                              "OtherData2":{
                                "type": "string"
                              }
                            },
                            "required": [
                              "CustomerName","PHONE","EMAIL","RFID","REG_NUM","VIN","FUEL_TYPE"
                            ]
                          }
                        }
                      }
                  },
              "responses": {
                "200": {
                  "description": "Successfully added the user",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                },
                "400": {
                  "description": "Bad request"
                }
              }
            }
          },
          "/api/registration/{userId}": {
            "get": {
              "description": "Get a user registred by passing the userId",
              "produces": [
                "application/json"
              ],
              "parameters": [
                {
                    "name": "userId",
                    "in": "path",
                    "description": "ID of the user",
                    "required": true,
                    "schema": {
                      "type": "string"
                    }
                  }
              ],
              "responses": {
                "200": {
                  "description": "Successfully got the user",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                }
              }
            },
            "delete": {
              "description": "Delete a user by passing the userId",
              "parameters": [
                {
                    "name": "userId",
                    "in": "path",
                    "description": "ID of the user",
                    "required": true,
                    "schema": {
                      "type": "string"
                    }
                  }
              ],
              "responses": {
                "200": {
                  "description": "Successfully deleted the user"
                }
              }
            }
          },
          "/api/registration/edit/{userId}": {
            "put": {
              "description": "Edit user details",
              "produces": [
                "application/json"
              ],
              "parameters": [
                {
                    "name": "userId",
                    "in": "path",
                    "description": "ID of the user",
                    "required": true,
                    "schema": {
                      "type": "string"
                    }
                  }
              ],
              "requestBody":{
                "description": "Edit user details",
                "content": {
                    "application/json":
                    {
                        "schema": {
                            "type": "object",
                            "properties": {
                              "CustomerName": {
                                "type": "string"
                              },
                              "PHONE":{
                                "type": "string"
                              },
                              "EMAIL":{
                                "type": "string"
                              },
                              "ADDRESS":{
                                "type": "string"
                              },
                              "RFID":{
                                "type": "string"
                              },
                              "REG_NUM":{
                                "type": "string"
                              },
                              "VIN":{
                                "type": "string"
                              },
                              "MAC":{
                                "type": "string"
                              },
                              "VEHICLE_MAKE":{
                                "type": "string"
                              },
                              "FUEL_TYPE":{
                                "type": "string"
                              },
                              "IS_OBD":{
                                "type": "string"
                              },
                              "DateOfCreation":{
                                "type": "string"
                              },
                              "DateOfUpdation":{
                                "type": "string"
                              },
                              "FUEL_LEVEL":{
                                "type": "string"
                              },
                              "TANK_CAPACITY":{
                                "type": "string"
                              },
                              "RPM":{
                                "type": "string"
                              },
                              "SPEED":{
                                "type": "string"
                              },
                              "MAF":{
                                "type": "string"
                              },
                              "COOLANT_TEMP":{
                                "type": "string"
                              },
                              "MILAGE":{
                                "type": "string"
                              },
                              "OtherData1":{
                                "type": "string"
                              },
                              "OtherData2":{
                                "type": "string"
                                }
                              }
                            },
                        "encoding": {
                          "payload": {
                            "contentType": "application/json"
                          }
                        }
                      }
                    }
                  },
              "responses": {
                "200": {
                  "description": "Successfully edited the user",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                },
                "400": {
                  "description": "Bad request"
                }
              }
            }
          },
          "/existing": {
            "post": {
              "description": "Hit this api for checking if the topic is already in the database and subscribing to it",
              "produces": [
                "application/json"
              ],
              "requestBody":{
                "description": "Post the registration number as the mqtt topic",
                "content": {
                    "application/json":
                    {
                        "schema": {
                            "type": "object",
                            "properties": {
                              "REG_NUM":{
                                "type": "string"
                              }
                            }
                          },
                        "encoding": {
                          "payload": {
                            "contentType": "application/json"
                          }
                        }
                      }
                    }
                  },
              "responses": {
                "200": {
                  "description": "Topic found and subscribed"
                }
              }
            }
          },
          "/existing/topics": {
            "get": {
              "description": "Get all the active topics",
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "Successfully got the subscribed topics",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                }
              }
            }
          },
          "/alerts": {
            "get": {
              "description": "Mqtt data is pushed from different active topics",
              "produces": [
                "application/json"
              ],
              "responses": {
                "200": {
                  "description": "Successfully got the alert message of respective topics",
                  "schema": {
                    "type": "array",
                    "items": [{}]
                  }
                }
              }
            }
          }    
    }
  }